+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Mutate Vintage Workload Cluster Apps to disable Pod Security Policies.
+
 ### Removed
 
 - Removed duplicate of Github pre-commit workflow.